Spent a week on and off playing with this to try and get new offsets for 4.0 to update Strainer as I had some time to kill :)
Currently I have Teleportation between stages working, teleporting to the boss if it's applicable to the current stage you're in. Health, mana, Gold, BossHP, CharacterID and Items should also be OK but only for Shovel of Hope.

I still need to put in some more time and find the root pointers for the full game for this to be finalised.

Committing this for the time being if anyone wanted the values to play with for themselves...